Refactored main.py into modularized files using cogs
Created utils.py for reusable functions accros the cogs, a cogs folder with all the reformatted commands in it and a data.py temporarily that will be replaces with a json/DB
Also did some general modification but this is a commit just to setup architecture/ideas/todos, neither race_game or the general_commands are done
