+++ data.py
@@ -1,0 +1,4 @@
+# TODO this is a temporary file to store data, will be replaced by a database/json file
+
+race_games = {}
+players = {}
